PR comments addressed Readme enhanced
