+++ UserSync/UserSync.csproj
@@ -226,8 +226,8 @@
     <Compile Include="Models\MSGraphUser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
-    <Compile Include="Utils\MSALMemoryTokenCache.cs" />
-    <Compile Include="Utils\TokenCacheHelper.cs" />
+    <Compile Include="Utils\MSALUserTokenMemoryCache.cs" />
+    <Compile Include="Utils\MSALAppTokenMemoryCache.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Content\bootstrap.css" />
